# subproyectoIntegracion-l8-2/src/main/webapp/docs/swagger.yaml
openapi: 3.0.0
servers:
  - url: 'https://city-events-api.ew.r.appspot.com/api'
  - url: 'http://localhost:8089/api'
info:
  description: Api about cities and events that take place in them
  version: 1.0.0
  title: CityEventsAPI
  contact:
    email: javnunrui@alum.us.es
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /fqas:
    get:
      tags:
        - fqas
      summary: Get fqas
      description: Get all fqas with the right attributes
      operationId: getFqas
      parameters:
        - name: limit
          in: query
          description: The maximum number of items to return
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: offset
          in: query
          description: The number of items to skip before starting to collect the result
          required: false
          schema:
            type: integer
            format: int32
            minimum: 0
        - name: q
          in: query
          description: Query to filter
          required: false
          schema:
            type: string
        - name: fields
          in: query
          description: Returns only the properties of the fqas
          required: false
          schema:
            type: string
        - name: order
          in: query
          description: Order results by question or answer
          required: false
          schema:
            type: string
            enum:
              - question
              - '-question'
              - answer
              - '-answer'
      responses:
        '200':
          description: Search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Fqa'
        '400':
          description: Bad request
    post:
      tags:
        - fqas
      summary: Create a fqa
      description: Create a fqa with the rights attributes
      operationId: createFqa
      requestBody:
        description: Fqa to create
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Fqa'
      responses:
        '201':
          description: Fqa created correctly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fqa'
        '400':
          description: Bad request
    put:
      tags:
        - fqas
      summary: Update a fqa
      description: Update a fqa introducing new values for certain attributes
      operationId: updateFqa
      requestBody:
        description: Fqa to update
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Fqa'
      responses:
        '204':
          description: No content. Fqa updated
        '404':
          description: Fqa not found
  '/fqas/{id}':
    get:
      tags:
        - fqas
      summary: Get a fqa by ID
      description: Get a fqa by ID with the right attributes
      operationId: getFqa
      parameters:
        - name: id
          in: path
          description: Id of the fqa
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Fqa with corresponding id found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Fqa'
        '404':
          description: Fqa not found
    delete:
      tags:
        - fqas
      summary: Delete a fqa
      description: Delete a fqa
      operationId: deleteFqa
      parameters:
        - name: id
          in: path
          description: Id of the fqa
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Fqa deleted
        '404':
          description: Fqa not found
components:
  schemas:
...
    Fqa:
      type: object
      required:
        - question
        - answer
      properties:
        id:
          type: string
        question:
          type: string
        answer:
          type: string
      example:
        id: f15
        question: Area for reduced mobility?
        answer: There is a conveniently equipped area for people with reduced mobility...